A spoonful of syntactic sugar
Obviously you started with java and you're used to its syntax, but if
you're using C# you may as well try some of its finer points, for
example, Automatic Properties, which I have added. See GameController,
where the getters have been replaced by public properties that are in
fact abstract getters and private setters. Finally, I've deleted a junk file, UpgradeLog.htm that VS/Mono creates when it encounters a project that needs to be upgraded and added it to the git ignore file.