Beginnings of the research selector
Yet to add the ability to set research, and have to resolve array out of
bounds errors that will arise, but should be easy enough to do tomorrow.
Once the research adding GUI has been made, we should be able to test
all the code in the game and hopefully fix it in a few days. Then it's
AI time. Finally, in case it is unclear the solution I have here is most
likely temporary, however it should be easy enough to duplicate it for
all startables and parts. I would estimate by Saturday the game will be
"playable"