Research Browser
Browser now displays next and previous properly, can deal with out of
bounds exceptions etc. In theory it should be able to star research as
well. Still to do, change the view when research is being done, maybe
make a meter or something that'll fill up. Enable the player to cancel
research to reaccess the browser. Then copy and paste all of this for
the, software, hardware and parts.