Fixed a Software Bug
Finish Software did not work properly, it will now. Added a method that
makes the whole picker reappear if it has been set to inactive, however
it will only go to the first page of the chapter it was last on.
Obviously if you click on research to see it in detail, then decide it
isn't for you, you don't want to go all the way back to page 1 if you
are on page 4 of 6. It would get annoying, I'll extend this method
later, but you can use it in your information window implementation for
now.