Fixed the Research picker
Problem lied in fields not being set back to null and my new methods for
possible research looking for the completed research when reentering the
Research section. This threw a nullReferenceException. This method was
made for reentering the section after completing research, not after
browsing other sections. Anyway, simple try catch and everything works
again.