Pretty much functional hierarchy screen, with camera controller inc. drag/zoom support.
Because we now have **TWO** (Yes sir, count 'em!) screens, I've added a camera controller, with a debug GO (GameObject)  that shows the bottom left/top right world co-ords. Currently it starts on the hierarchy screen (which was for testing and is obviously otherwise wrong) and has no functionality regarding going to the main screen (you can just move the camera manually over the main screen if you are on this commit.

[BUG] Max zoom out causes some serious shit to occur